requirements.txt
fastapi 
uvicorn 
sentence-transformers 
faiss-cpu 
openai 
python-dotenv
langchain
langchain-openai
langchain-groq